test(#664): relocate 5 real-process boundary suites to integration (C6.1-spawn)
file-lock (spawns a bun child to prove process.exit fires the exit-handler
lock release), index-writer-lock (real PID/lock subprocess), registry-build-index
+ tar-utils-scan (real `tar` archive create/extract), and github (imports
node:child_process to spy on the `gh auth token` spawn) all cross a real process
boundary at the unit-lint level. Relocate out of the unit glob into
tests/integration/ (fix import depth) and drop from ALLOWED_SPAWN, lowering
UNIT_PURITY_BASELINE 28 -> 23.
